# Allow Buffer.getByte() to search backwards for the desired segment