Uses of Buffer.getByte() within the Okio project access it more often relative to the end of the Buffer instead of from the beginning.  Specifically the ReadBufferedSource.readHexadecimalUnsignedLong() and ReadBufferedSource.readDecimalLong() methods make heavy use of retrieving the latest byte from the Buffer and with this change, the unit tests for these methods with the OneByteAtATime source no longer take over a second to run.

A benchmark for this behavior has been created and verified to show significant improvement.

Before:
```
Benchmark                                    (bufferSize)    Mode  Samples  Score   Error  Units
c.s.o.b.GetByteBenchmark.getByteBeginning         2097152  sample   160460  0.037 ± 0.003  us/op
c.s.o.b.GetByteBenchmark.getByteEnd               2097152  sample   107573  1.481 ± 0.018  us/op
c.s.o.b.GetByteBenchmark.getByteMiddle            2097152  sample   112678  0.721 ± 0.013  us/op
```

After:
```
Benchmark                                    (bufferSize)    Mode  Samples  Score   Error  Units
c.s.o.b.GetByteBenchmark.getByteBeginning         2097152  sample   167078  0.035 ± 0.003  us/op
c.s.o.b.GetByteBenchmark.getByteEnd               2097152  sample   136939  0.036 ± 0.003  us/op
c.s.o.b.GetByteBenchmark.getByteMiddle            2097152  sample   120560  0.672 ± 0.015  us/op
```

In total, during all unit tests, the getByte() method is invoked 87,467 times.  Roughly 60% (52,152) of those are for bytes in the second half of the buffer.  Without the change, this results in an additional 67,196,832 passes through the segment search loop.  With the change, there are only an additional 6,213 passes through the forward loop and an additional 10,281 passes through the backwards loop.