@swankjesse 2018-02-03
Nice PR.

For `select()`, I don’t think the unit tests are necessarily representative of real-world use. I spent a lot of time trying to tune `select()` to be faster for JSON parsing. I wrote lots of fancy code inside Options to precompute stuff to save work if options contained common prefixes.

In the end the naive brute force won. How frustrating.

If you’d like to take another run at it, please do. The thing I think we’re optimizing here is downstream users of select(). Most notably, JSON parsing in Moshi. In such cases the number of options is typically the number of fields of a bound object; typically between 2 and 10 and rarely greater than 20.

The API design is intended to permit better optimizations eventually, and perhaps even ones that are asymmetric. If you can find something that’s really fast but only works up to 4 options, that’s allowed!